requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9

# Supabase (Task 2)
supabase==2.9.0
postgrest==0.17.2
httpx==0.27.0

# Google Cloud Vision (Task 3)
google-cloud-vision==3.7.2

# OpenAI (Task 3)
openai==1.51.0

# Image Processing (Task 4)
pillow==10.3.0

# Le seguenti dipendenze verranno aggiunte nei prossimi task:
# Task 2 - Supabase
# supabase==2.3.0

# Task 3 - Google Cloud Vision
# google-cloud-vision==3.5.0

# Task 3 - OpenAI
# openai==1.6.1

# Task 4 - Image Processing
# pillow==10.1.0

# Development (opzionali)
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.12.1